Full project with both parts (Read description)
I merged both parts, the first one should be done, the second one is not!
(Note that I added the code for d) and e) for the second part.)

Problems with the second part:
1. If the charge is negative (comment out the np.random.seed(14), the things don't work anymore. Somehow the functions circular_path_coefficients() and circular_path_center_coordinates() still need to be changed. Also, sometimes the paths need to go up/outwards, this should still be implemented.
2. Also, when having 100+ trajectories, some trajectories are not drawn completely (some points of the end of the magnetic field are missing).

I think it's better to solve 1. first because the second one also depends on the angle and stuff, so this first needs to be correct to tell how far the points can go.