New Version, only few things remaining!
This is an improved version of the code (Thanks @mariano!), negative charge should work, lengths of the trajectories should work!

Only thing remaining is that now and then there are strange curves in parts where they should not be (see WhatsApp image). It would be great if anyone finds a solution!